chore: Make core/ and move non-background, non-content-scripts there
