+++ extension/core/mediaItem.js
